requirements.txt
Django==3.2.10
django-confy==1.0.4
django-reversion==4.0.1
django-extensions==3.1.5
psycopg2==-2.9.3
djangorestframework==3.13.1



# sudo apt install gdal-bin
